heating.tingapp/requirements.txt
paho-mqtt
tingbot-python>=1.0.2
